+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Unreleased
 
+* Add `Lang::is_machine_translation`.
+    - Adds support for `machine` variant, as in `en-machine` to indicate a resource that was machine translated.
+    - `LangMap` gives priority for human translations of different regions over machine translations.
+
+* Improve `cargo zng l10n --check`.
+    - Check the Fluent syntax of localizations.
+    - Check if localized messages match the template.
+    - Add `--check-strict` to also check for missing messages.
+
+* Add machine translated l10n resources for: ja, es.
 
 # 0.21.3